Разделила код на модули, подключила главный модуль
