+++ js/comments.js
@@ -1,0 +1,5 @@
+const entryComment = 'Строка ввода комментария'; //input.value;
+
+const getLengthLine = (line, maxLine) => line.length <= maxLine;
+
+getLengthLine(entryComment, 140);
